Add prime number logic
To determine whether a number T is prime or not the an algorithm can be 
used. 

Take the square root of the number T and round up to nearest integer to 
get N. For every M number in the range 2 to N, divide the number T being 
tested by M, and if the number is an integer as opposed to a float, then 
T is not a prime number. But if none of the results are integers then 
the number T is a prime number.