Add add_new_stock function to Inventory class
The inventory system should hold a stocks dictionary object with the 
add_new_stock function creating a new index key. The key should be 
assigned a dictionary value including the price and quantity.

The total_items attribute of the Inventory instance should also be 
updated.

Use a fixture to create an Inventory object for each test and prevent 
repeated code.