Merge add_new_stock function tests
The add_new_stock tests can be merged, with the 'success' test added to 
the bad input test by updating the parametrized function. The else 
condition of the try block can then be used to assert a successful stock 
addition.

Fixtures can be used with parametrized functions, where it just needs to 
be included as a test function parameter and not in the parametrize 
decorator