Add remove_stock function to Inventory class
A remove_stock function is required to remove existing quantities of 
existing items held in stocks. It should take a name and quantity as 
arguements.

Exceptions should be raised if the item key does not exist in stocks, if 
the quantity is specified is less than 1 or if the quantity is more than 
exists in stocks.

If valid arguements are passed the inventory stocks and total_items 
attributes should be updated.